Avoid ClassCastException with Int/Long headers
When a Date based header (eg Last Modified) is added to the PageInfo it is treated as a Header<Long>. Header's getValue() method is generically typed - returning Long in this case. The former code tried to cast Long --> String which fails with a ClassCastException. The current header types (String, Long, Int) shouldn't need anything more sophisticated than a toString() conversion to avoid this issue.